requirements.txt
langchain
langchain-community
langchain-core
python-dotenv
fastapi
uvicorn
pinecone
langchain_pinecone
langmem
langchain-postgres
langchain-openai==0.3.32
-e .








